+++ server/map_rpc_server.go
@@ -53,6 +53,10 @@ type TrillianMapServerOptions struct {
 	// UseSingleTransaction specifies whether updates to a map should be
 	// attempted within a single transaction.
 	UseSingleTransaction bool
+
+	// UseLargePreload enables the performance workaround applied when
+	// UseSingleTransaction is set.
+	UseLargePreload bool
 }
 
 // TrillianMapServer implements the RPC API defined in the proto
@@ -263,6 +267,33 @@ func (t *TrillianMapServer) SetLeaves(ctx context.Context, req *trillian.SetMapL
 				HashedValue: l.LeafHash,
 			})
 		}
+
+		// Work around a performance issue when using the map in
+		// single-transaction mode by preloading all the nodes we know the
+		// sparse merkle writer is going to need.
+		if t.opts.UseSingleTransaction && t.opts.UseLargePreload {
+			readRev, err := tx.ReadRevision(ctx)
+			if err != nil {
+				return err
+			}
+			nidSet := make(map[string]bool)
+			nids := make([]storage.NodeID, 0, len(hkv)*256)
+			for _, i := range hkv {
+				nid := storage.NewNodeIDFromHash(i.HashedKey)
+				sibs := (&nid).Siblings()
+				for _, sib := range sibs {
+					sibID := sib.String()
+					if _, ok := nidSet[sibID]; !ok {
+						nidSet[sibID] = true
+						nids = append(nids, sib)
+					}
+				}
+			}
+			if _, err := tx.GetMerkleNodes(ctx, readRev, nids); err != nil {
+				return err
+			}
+		}
+
 		if err = smtWriter.SetLeaves(ctx, hkv); err != nil {
 			return err
 		}