+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 Not yet released; provisionally v2.0.0 (may change).
 
+### Add Profiling Flags to Binaries
+
+The `trillian_log_server`, `trillian_log_signer` and `trillian_map_server`
+binaries now have CPU and heap profiling flags. Profiling is off by default.
+For more details see the
+[Go Blog](https://blog.golang.org/profiling-go-programs).
+### Map performance tweaks
+
+The map mode has had some performance tweaks added:
+* A workaround for locking issues which affect the map when it's used in
+  single-transaction mode.
+
 ### Introduce BatchInclusionProof function
 
 Added a batch version of the Merkle Tree InclusionProof function.
@@ -178,6 +190,12 @@ The `SubtreeCache.GetNodeHash()` method is no longer exported.
 
 The memory storage provider has been refactored to make it more consistent with
 the other storage providers.
+
+The `LogMetadata.GetUnsequencedCounts()` method has been removed.
+
+`NodeReader.GetMerkleNodes` now must return `Node` objects in the same order as
+node IDs requested. Storage implementations known to us already adhere to this
+requirement.
 
 ### Maphammer improvements